# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Marketplace de Servicios API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: List active job postings
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: locality
          in: query
          schema:
            type: ["string", "null"]
        - name: industry
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of job postings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobPostingWithCompany"

    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job posting (company only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPosting"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job posting detail with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPostingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPosting"

    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /jobs/{id}/apply:
    post:
      operationId: applyToJob
      tags: [jobs]
      summary: Apply to a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyJobBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplication"
        "400":
          description: Already applied or error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/applications:
    get:
      operationId: getJobApplications
      tags: [jobs]
      summary: Get applications for a job (company only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobApplicationWithDetails"
  /jobs/applications/{id}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [jobs]
      summary: Update application status (company only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationStatusBody"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplication"
  /cvs/public:
    get:
      operationId: getPublicCvs
      tags: [cvs]
      summary: Get public CVs (company only)
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: locality
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of public CVs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicCv"
...
  /admin/companies/{id}/approve:
    patch:
      operationId: adminApproveCompany
      tags: [admin]
      summary: Approve or reject a company account (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                approved:
                  type: boolean
              required:
                - approved
      responses:
        "200":
          description: Company status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    JobPosting:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        title:
          type: string
        description:
          type: string
        industry:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
        modality:
          type: string
          enum: [presencial, remoto, hibrido]
        contractType:
          type: string
          enum: [full_time, part_time, freelance, pasantia]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        requirements:
          type: ["string", "null"]
        benefits:
          type: ["string", "null"]
        isActive:
          type: boolean
        adminApproved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - companyId
        - title
        - description
        - modality
        - contractType
        - isActive
        - adminApproved
        - createdAt
    JobQuestion:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
          enum: [text, single_choice, multiple_choice]
        options:
          type: ["array", "null"]
          items:
            type: string
        required:
          type: boolean
        sortOrder:
          type: integer
      required:
        - id
        - jobId
        - questionText
        - questionType
        - required
        - sortOrder
    JobCompanySummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        companyName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
        companyIndustry:
          type: ["string", "null"]
      required:
        - id
        - name
    JobPostingWithCompany:
      allOf:
        - $ref: "#/components/schemas/JobPosting"
        - type: object
          properties:
            company:
              $ref: "#/components/schemas/JobCompanySummary"
          required:
            - company
    JobPostingDetail:
      allOf:
        - $ref: "#/components/schemas/JobPosting"
        - type: object
          properties:
            company:
              $ref: "#/components/schemas/JobCompanySummary"
            questions:
              type: array
              items:
                $ref: "#/components/schemas/JobQuestion"
          required:
            - company
            - questions
    CreateJobBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        industry:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
        modality:
          type: string
          enum: [presencial, remoto, hibrido]
        contractType:
          type: string
          enum: [full_time, part_time, freelance, pasantia]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        requirements:
          type: ["string", "null"]
        benefits:
          type: ["string", "null"]
        questions:
          type: array
          items:
            type: object
            properties:
              questionText:
                type: string
              questionType:
                type: string
                enum: [text, single_choice, multiple_choice]
              options:
                type: ["array", "null"]
                items:
                  type: string
              required:
                type: boolean
            required:
              - questionText
      required:
        - title
        - description
    UpdateJobBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
        modality:
          type: ["string", "null"]
          enum: [presencial, remoto, hibrido, null]
        contractType:
          type: ["string", "null"]
          enum: [full_time, part_time, freelance, pasantia, null]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        requirements:
          type: ["string", "null"]
        benefits:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        questions:
          type: ["array", "null"]
          items:
            type: object
            properties:
              questionText:
                type: string
              questionType:
                type: string
                enum: [text, single_choice, multiple_choice]
              options:
                type: ["array", "null"]
                items:
                  type: string
              required:
                type: boolean
            required:
              - questionText
    JobApplication:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        applicantId:
          type: integer
        coverLetter:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, reviewed, accepted, rejected]
        createdAt:
          type: string
      required:
        - id
        - jobId
        - applicantId
        - status
        - createdAt
    ApplyJobBody:
      type: object
      properties:
        coverLetter:
          type: ["string", "null"]
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              answerText:
                type: ["string", "null"]
            required:
              - questionId
    JobAnswer:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        questionId:
          type: integer
        answerText:
          type: ["string", "null"]
        question:
          $ref: "#/components/schemas/JobQuestion"
      required:
        - id
        - applicationId
        - questionId
    JobApplicantSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
        cvUrl:
          type: ["string", "null"]
      required:
        - id
        - name
        - email
    JobApplicationWithDetails:
      allOf:
        - $ref: "#/components/schemas/JobApplication"
        - type: object
          properties:
            applicant:
              $ref: "#/components/schemas/JobApplicantSummary"
            answers:
              type: array
              items:
                $ref: "#/components/schemas/JobAnswer"
          required:
            - applicant
            - answers
...
    UpdateApplicationStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, reviewed, accepted, rejected]
      required:
        - status
    PublicCv:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
        cvUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - createdAt
...