# lib/api-zod/src/generated/types/applyJobBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketplace de Servicios API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplyJobBodyAnswersItem } from "./applyJobBodyAnswersItem";

export interface ApplyJobBody {
  /** @nullable */
  coverLetter?: string | null;
  answers?: ApplyJobBodyAnswersItem[];
}