# lib/api-zod/src/generated/types/jobQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketplace de Servicios API
 * OpenAPI spec version: 0.1.0
 */
import type { JobQuestionQuestionType } from "./jobQuestionQuestionType";

export interface JobQuestion {
  id: number;
  jobId: number;
  questionText: string;
  questionType: JobQuestionQuestionType;
  /** @nullable */
  options?: string[] | null;
  required: boolean;
  sortOrder: number;
}