# lib/api-zod/src/generated/types/updateJobBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketplace de Servicios API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateJobBodyContractType } from "./updateJobBodyContractType";
import type { UpdateJobBodyModality } from "./updateJobBodyModality";
import type { UpdateJobBodyQuestionsItem } from "./updateJobBodyQuestionsItem";

export interface UpdateJobBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  locality?: string | null;
  /** @nullable */
  modality?: UpdateJobBodyModality;
  /** @nullable */
  contractType?: UpdateJobBodyContractType;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  /** @nullable */
  requirements?: string | null;
  /** @nullable */
  benefits?: string | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  questions?: UpdateJobBodyQuestionsItem[] | null;
}